requirements.txt
Faker==13.3.4
python-dotenv==1.0.1